Mini_Capstone/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pymongo>=4.8,<5.0
python-jose[cryptography]>=3.3,<4.0
passlib>=1.7,<2.0
pydantic-settings>=2.5,<3.0
email-validator>=2.2,<3.0
python-multipart>=0.0.9,<1.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
